[refactor](home): 홈 TopAppBar 에 있는 Image -> Icon 으로 변경 및 common 으로 이동 (#13)
