+++ app/src/main/java/com/kuit/ourmenu/ui/common/OurMenuLogoTopAppBarWithAddButton.kt
@@ -1,0 +1,51 @@
+package com.kuit.ourmenu.ui.common
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.kuit.ourmenu.R
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun OurMenuLogoTopAppBarWithAddButton(modifier: Modifier = Modifier) {
+    // 기본
+    TopAppBar(
+        modifier = Modifier
+            .fillMaxWidth(),
+        title = {
+            Icon(
+                painter = painterResource(R.drawable.ic_ourmenu_text_logo),
+                contentDescription = "Top App Bar Logo",
+                modifier = Modifier.padding(start = 8.dp),
+                tint = Color.Unspecified
+            )
+        },
+        actions = {
+            IconButton(
+                onClick = { /* TODO : Add Menu Button Click Event */ },
+                modifier = Modifier.padding(end = 20.dp)
+            ) {
+                Icon(
+                    painter = painterResource(R.drawable.ic_home_plus),
+                    contentDescription = "Top App Bar Add Menu Button",
+                    tint = Color.Unspecified
+                )
+            }
+        }
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun HomeTopAppBarPreview() {
+    OurMenuLogoTopAppBarWithAddButton()
+}
