+++ app/src/main/java/com/kuit/ourmenu/ui/home/screen/HomeScreen.kt
@@ -1,10 +1,8 @@
 package com.kuit.ourmenu.ui.home.screen
 
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
@@ -15,7 +13,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.kuit.ourmenu.ui.home.component.recommendation.sub.HomeSubRecommendation
-import com.kuit.ourmenu.ui.home.component.HomeTopAppBar
+import com.kuit.ourmenu.ui.common.OurMenuLogoTopAppBarWithAddButton
 import com.kuit.ourmenu.ui.home.component.recommendation.main.HomeMainRecommendation
 import com.kuit.ourmenu.ui.home.dummy.HomeDummyData
 
@@ -26,7 +24,7 @@ fun HomeScreen(modifier: Modifier = Modifier) {
 
     Scaffold(
         topBar = {
-            HomeTopAppBar()
+            OurMenuLogoTopAppBarWithAddButton()
         }
     ) { innerPadding ->
         Column(
